+++ csrc/dynamic_transform.cpp
@@ -169,7 +169,7 @@ class DynamicTransformInitialInfoBuilder : public IterVisitor {
         TensorDomain::noReductions(inp_tv->getMaybeRFactorDomain());
     const std::vector<IterDomain*>& out_root = out_tv->getRootDomain();
     NVF_ERROR(inp_rfactor.size() == out_root.size());
-    for (auto i : c10::irange(out_root.size())) {
+    for (auto i : c10::irange((int64_t)out_root.size())) {
       IterDomain* out_id = out_root[i];
       if (!out_id->isSymbolic()) {
         continue;
@@ -221,7 +221,7 @@ class DynamicTransformInitialInfoBuilder : public IterVisitor {
     // Vals. These will be the inputs that are explicitly used in the cache ID
     // for KernelArgumentHolder.
     auto dyn_vals = info_.getRootDynamicVals();
-    for (const auto i : c10::irange(info_.fusion()->inputs().size())) {
+    for (const auto i : c10::irange((int64_t)info_.fusion()->inputs().size())) {
       auto input = info_.fusion()->inputs().at(i);
       if (dyn_vals.find(input) != dyn_vals.end()) {
         info_.scalar_inputs_affecting_concretization_.insert(i);
@@ -272,7 +272,7 @@ DynamicTransformConcretizationInfo::DynamicTransformConcretizationInfo(
   analyzeFactoryOutputs(expr_eval);
 
   auto maybe_zero_extents = initial_info_->getMaybeZeroExtents();
-  for (auto i : c10::irange(maybe_zero_extents.size())) {
+  for (auto i : c10::irange((int64_t)maybe_zero_extents.size())) {
     auto ext = maybe_zero_extents.at(i);
     auto ext_opt = expr_eval->evaluate(ext);
     NVF_ERROR(
@@ -288,7 +288,7 @@ DynamicTransformConcretizationInfo::DynamicTransformConcretizationInfo(
 void DynamicTransformConcretizationInfo::analyzeReshapes(
     ExpressionEvaluator* expr_eval) {
   const auto& reshape_tvs = initial_info_->getDynamicReshapedTensorViews();
-  for (const auto tv_index : c10::irange(reshape_tvs.size())) {
+  for (const auto tv_index : c10::irange((int64_t)reshape_tvs.size())) {
     auto out_tv = reshape_tvs.at(tv_index);
     auto op = out_tv->definition()->as<ViewOp>();
     auto inp_tv = op->in()->as<TensorView>();
@@ -308,7 +308,7 @@ void DynamicTransformConcretizationInfo::analyzeReshapes(
 
     // Determine input shape using expr evaluator
     std::vector<int64_t> inp_shape(inp_dom.size(), 0);
-    for (const auto i : c10::irange(inp_dom.size())) {
+    for (const auto i : c10::irange((int64_t)inp_dom.size())) {
       auto inp_id = inp_dom.at(i);
       // This should have been validated when initially creating reshape
       // op, but just in case
@@ -337,7 +337,7 @@ void DynamicTransformConcretizationInfo::analyzeReshapes(
     // Determine output shape using expr evaluator. Note there may be
     // one domain of extent -1
     std::vector<int64_t> out_shape(out_dom.size(), 0);
-    for (const auto i : c10::irange(out_dom.size())) {
+    for (const auto i : c10::irange((int64_t)out_dom.size())) {
       auto out_id = out_dom.at(i);
       auto extent_val = expr_eval->evaluate(out_id->extent());
       NVF_ERROR(
@@ -367,7 +367,7 @@ void DynamicTransformConcretizationInfo::analyzeReshapes(
 void DynamicTransformConcretizationInfo::analyzeResizes(
     ExpressionEvaluator* expr_eval) {
   const auto& resize_ids = initial_info_->getDynamicResizedIterDomains();
-  for (const auto id_index : c10::irange(resize_ids.size())) {
+  for (const auto id_index : c10::irange((int64_t)resize_ids.size())) {
     auto out_id = resize_ids.at(id_index);
     auto op = out_id->definition()->as<Resize>();
 
@@ -404,7 +404,7 @@ void DynamicTransformConcretizationInfo::analyzeExpands(
     ExpressionEvaluator* expr_eval) {
   const std::vector<TensorView*>& expanded_tvs =
       initial_info_->getDynamicExpandedTensorViews();
-  for (const auto tv_index : c10::irange(expanded_tvs.size())) {
+  for (const auto tv_index : c10::irange((int64_t)expanded_tvs.size())) {
     const TensorView* out_tv = expanded_tvs.at(tv_index);
     const TensorView* inp_tv = out_tv->definition()->as<ExpandOp>()->in();
 
@@ -452,11 +452,11 @@ void DynamicTransformConcretizationInfo::analyzeFactoryOutputs(
   const std::vector<TensorView*>& factory_tvs =
       initial_info_->getDynamicFactoryOutputs();
   factory_output_itertypes_.reserve(factory_tvs.size());
-  for (const auto tv_index : c10::irange(factory_tvs.size())) {
+  for (const auto tv_index : c10::irange((int64_t)factory_tvs.size())) {
     const TensorView* tv = factory_tvs.at(tv_index);
     const std::vector<IterDomain*>& rf = tv->getMaybeRFactorDomain();
     std::vector<std::pair<int64_t, IterType>> conc_iter_types;
-    for (int64_t pos : c10::irange(rf.size())) {
+    for (int64_t pos : c10::irange((int64_t)rf.size())) {
       const IterDomain* id = rf[pos];
       if (!id->isSymbolic()) {
         continue;
@@ -492,7 +492,7 @@ bool DynamicTransformConcretizationInfo::operator==(
     return false;
   }
 
-  for (const auto i : c10::irange(reshape_transforms_.size())) {
+  for (const auto i : c10::irange((int64_t)reshape_transforms_.size())) {
     const auto& analysis = reshape_transforms_.at(i);
     const auto& other_analysis = other.reshape_transforms_.at(i);
     if (analysis != other_analysis) {
@@ -500,7 +500,7 @@ bool DynamicTransformConcretizationInfo::operator==(
     }
   }
 
-  for (const auto i : c10::irange(resize_itertypes_.size())) {
+  for (const auto i : c10::irange((int64_t)resize_itertypes_.size())) {
     const auto& itertype = resize_itertypes_.at(i);
     const auto& other_itertype = other.resize_itertypes_.at(i);
     if (itertype != other_itertype) {
@@ -512,7 +512,7 @@ bool DynamicTransformConcretizationInfo::operator==(
     return false;
   }
 
-  for (const auto i : c10::irange(expand_axes_.size())) {
+  for (const auto i : c10::irange((int64_t)expand_axes_.size())) {
     const auto& expand_axes = expand_axes_.at(i);
     const auto& other_expand_axes = other.expand_axes_.at(i);
     if (expand_axes != other_expand_axes) {
@@ -520,7 +520,7 @@ bool DynamicTransformConcretizationInfo::operator==(
     }
   }
 
-  for (const auto i : c10::irange(empty_extents_.size())) {
+  for (const auto i : c10::irange((int64_t)empty_extents_.size())) {
     const auto& ee = empty_extents_.at(i);
     const auto& other_ee = other.empty_extents_.at(i);
     if (ee != other_ee) {
@@ -580,7 +580,7 @@ std::string DynamicTransformConcretizationInfo::toString() const {
   NVF_ERROR(
       factory_output_itertypes_.size() ==
       initial_info_->getDynamicFactoryOutputs().size());
-  for (int64_t i : c10::irange(factory_output_itertypes_.size())) {
+  for (int64_t i : c10::irange((int64_t)factory_output_itertypes_.size())) {
     TensorView* tv = initial_info_->getDynamicFactoryOutputs().at(i);
     ss << indent << indent << tv->toString() << std::endl;
     for (const auto& [pos, iter_type] : factory_output_itertypes_.at(i)) {
@@ -779,7 +779,7 @@ void DynamicTransformConcretizer::concretizeReshape() {
     NVF_ERROR(
         old_rfactor.size() == new_rfactor.size(),
         "Concretized reshape rfactor size does not match symbolic rfactor");
-    for (auto idx : c10::irange(new_rfactor.size())) {
+    for (auto idx : c10::irange((int64_t)new_rfactor.size())) {
       auto old_extent = old_rfactor.at(idx)->extent();
       auto new_extent = new_rfactor.at(idx)->extent();
       // If the old extent did not have a definition, we don't need to replace
@@ -844,7 +844,7 @@ void DynamicTransformConcretizer::concretizeExpand() {
     std::vector<IterDomain*> out_rfactor =
         TensorDomain::noReductions(symbolic_out_tv->getMaybeRFactorDomain());
     NVF_ERROR(axis_is_expanded.size() == out_rfactor.size());
-    for (int64_t i : c10::irange(out_rfactor.size())) {
+    for (int64_t i : c10::irange((int64_t)out_rfactor.size())) {
       if (!axis_is_expanded[i]) {
         // Propagate as usual for non-expanded IterDomains
         continue;
@@ -870,7 +870,7 @@ void DynamicTransformConcretizer::concretizeFactoryOutputs() {
       info_->initialInfo()->getDynamicFactoryOutputs();
   const auto& pair_vecs = info_->getFactoryOutputIterTypes();
   NVF_ERROR(factory_tvs.size() == pair_vecs.size());
-  for (const int64_t i : c10::irange(factory_tvs.size())) {
+  for (const int64_t i : c10::irange((int64_t)factory_tvs.size())) {
     TensorView* tv = factory_tvs[i];
     const std::vector<std::pair<int64_t, IterType>>& pair_vec = pair_vecs[i];
     for (auto& [pos, iter_type] : pair_vec) {
@@ -964,7 +964,7 @@ void DynamicTransformConcretizer::mutate(TensorView* tv) {
       IterType iter_type = IterType::Symbolic;
       const auto input_ids =
           ir_utils::filterByType<IterDomain>(expr->inputs()).vector();
-      for (auto i : c10::irange(input_ids.size())) {
+      for (auto i : c10::irange((int64_t)input_ids.size())) {
         auto inp_id = input_ids.at(i);
         auto updated_id = maybeMutated(inp_id)->as<IterDomain>();
         NVF_CHECK(
@@ -1078,7 +1078,7 @@ void DynamicTransformConcretizer::mutate(TensorDomain* td) {
       new_maybe_alloc.size() == original_alloc.size(),
       "rank of allocation domain shouldn't change in concretization");
 
-  for (const auto i : c10::irange(original_alloc.size())) {
+  for (const auto i : c10::irange((int64_t)original_alloc.size())) {
     auto original_id = original_alloc.at(i);
     if (original_id->getIterType() != IterType::Symbolic) {
       continue;
@@ -1218,7 +1218,7 @@ bool DynamicTransformConcretizer::propagateFromProducerToConsumer(
 
   bool is_concretized = false;
 
-  for (const auto i : c10::irange(root_domain.size())) {
+  for (const auto i : c10::irange((int64_t)root_domain.size())) {
     auto root_id = root_domain.at(i);
     if (root_id->getIterType() != IterType::Symbolic) {
       continue;
