+++ csrc/scheduler/mma_utils.cpp
@@ -37,7 +37,7 @@ inline mma_utils::MmaDataTypes getMmaDataTypes(
 }
 
 //! Return sizes of smem_a, smem_b, smem_c in bytes
-std::tuple<size_t, size_t, size_t> computeSharedMemorySizes(
+std::tuple<int64_t, int64_t, int64_t> computeSharedMemorySizes(
     const MatMulTileOptions& gemm_tile,
     const MatmulParams::DoubleBufferOptions& double_buffer_options,
     const MmaDataTypes& data_types) {
@@ -45,24 +45,22 @@ std::tuple<size_t, size_t, size_t> computeSharedMemorySizes(
 
   auto warp_dims = gemm_tile.cta_tile / gemm_tile.warp_tile;
 
-  int ab_factor = double_buffer_options.double_buffer_smem_write
+  int64_t ab_factor = double_buffer_options.double_buffer_smem_write
       ? double_buffer_options.smem_double_buffer_stage
       : 1;
 
   // see scheduleContiguousVectorLoad
-  const int vector_word = 8;
-  const int round_to_factor = warp_dims.m * warp_dims.n * warp_dims.k *
+  const int64_t vector_word = 8;
+  const int64_t round_to_factor = warp_dims.m * warp_dims.n * warp_dims.k *
       properties->warpSize * vector_word;
-  const int mk = gemm_tile.cta_tile.m * gemm_tile.cta_tile.k;
-  const int nk = gemm_tile.cta_tile.n * gemm_tile.cta_tile.k;
-  const size_t smem_a =
-      (size_t)(ceilDiv(mk, round_to_factor) * round_to_factor * ab_factor) *
-      dataTypeSize(data_types[0]);
-  const size_t smem_b =
-      (size_t)(ceilDiv(nk, round_to_factor) * round_to_factor * ab_factor) *
-      dataTypeSize(data_types[1]);
-  const size_t smem_c = (size_t)(gemm_tile.cta_tile.m * gemm_tile.cta_tile.n) *
-      dataTypeSize(data_types[2]);
+  const int64_t mk = gemm_tile.cta_tile.m * gemm_tile.cta_tile.k;
+  const int64_t nk = gemm_tile.cta_tile.n * gemm_tile.cta_tile.k;
+  const int64_t smem_a = ceilDiv(mk, round_to_factor) * round_to_factor *
+      ab_factor * dataTypeSize(data_types[0]);
+  const int64_t smem_b = ceilDiv(nk, round_to_factor) * round_to_factor *
+      ab_factor * dataTypeSize(data_types[1]);
+  const int64_t smem_c =
+      gemm_tile.cta_tile.m * gemm_tile.cta_tile.n * dataTypeSize(data_types[2]);
 
   return {smem_a, smem_b, smem_c};
 }
@@ -270,7 +268,7 @@ void scheduleWarpTileWithNoReduction(TensorView* tv, MatMulTileOptions tile) {
       cta_tile.k % warp_tile.k == 0,
       "Number of warp on k dimension need to be integer");
 
-  int num_warp_k = cta_tile.k / warp_tile.k;
+  int64_t num_warp_k = cta_tile.k / warp_tile.k;
 
   //        -2  -1
   //[...    M,   N]
@@ -305,10 +303,10 @@ void scheduleWarpTileWithNoReduction(TensorView* tv, MatMulTileOptions tile) {
 void scheduleContiguousVectorLoad(
     TensorView* tv,
     MatMulTileOptions tile,
-    int vector_word,
+    int64_t vector_word,
     bool vectorize) {
   auto warp_dims = tile.cta_tile / tile.warp_tile;
-  int num_of_thread = warp_dims.m * warp_dims.n * warp_dims.k * 32;
+  int64_t num_of_thread = warp_dims.m * warp_dims.n * warp_dims.k * 32;
 
   tv->split(-1, num_of_thread * vector_word);
   tv->split(-1, vector_word);
